requirements.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.3.0
flower==0.9.4
humanize==0.5.1
importlib-metadata==1.6.1
kombu==4.6.10
pymongo==3.10.1
python-dateutil==2.6.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
six==1.15.0
tornado==6.0.4
vine==1.3.0
zipp==3.1.0
progressbar==2.5
